third_party/blink/web_tests/virtual/computed-width-values-depend-on-style-values/external/wpt/css/css-typed-om/the-stylepropertymap/properties/outline-width-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'outline-width' to the 'thin' keyword: thin
  assert_approx_equals: expected 1 +/- 0.000001 but got 0
[FAIL] Can set 'outline-width' to the 'medium' keyword: medium
  assert_approx_equals: expected 3 +/- 0.000001 but got 0
[FAIL] Can set 'outline-width' to the 'thick' keyword: thick
  assert_approx_equals: expected 5 +/- 0.000001 but got 0
Harness: the test ran to completion.
